+++ Hearthstone Deck Tracker/HsReplay/HSReplayNetHelper.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Hearthstone_Deck_Tracker.Controls.Error;
+using Hearthstone_Deck_Tracker.Hearthstone;
 using Hearthstone_Deck_Tracker.Utility;
 using Hearthstone_Deck_Tracker.Utility.Extensions;
 
@@ -8,6 +9,13 @@ namespace Hearthstone_Deck_Tracker.HsReplay
 {
 	internal static class HSReplayNetHelper
 	{
+		static HSReplayNetHelper()
+		{
+			ConfigWrapper.CollectionSyncingChanged += () => SyncCollection().Forget();
+			CollectionHelper.OnCollectionChanged += () => SyncCollection().Forget();
+		}
+
+		public static event Action CollectionSynced;
 		public static event Action<bool> Authenticating;
 
 		public static async Task TryAuthenticate()
@@ -33,5 +41,28 @@ namespace Hearthstone_Deck_Tracker.HsReplay
 			if(Account.Instance.TokenClaimed == false)
 				await HSReplayNetOAuth.ClaimUploadToken(Account.Instance.UploadToken);
 		}
+
+		public static async Task SyncCollection()
+		{
+			if(!Config.Instance.SyncCollection || !HSReplayNetOAuth.IsAuthenticated)
+				return;
+			var collection = await CollectionHelper.GetCollection();
+			if(collection == null)
+				return;
+			var hash = collection.GetHashCode();
+			var account = collection.AccountHi + "-" + collection.AccountLo;
+			if(Account.Instance.CollectionState.TryGetValue(account, out var state) && state.Hash == hash)
+				return;
+			if(await HSReplayNetOAuth.UpdateCollection(collection))
+			{
+				Account.Instance.CollectionState[account] = new Account.SyncState(hash);
+				Account.Save();
+				CollectionSynced?.Invoke();
+			}
+			else
+				ErrorManager.AddError("HSReplay.net error", "Could not update your collection. Please try again later.");
+		}
+
+
 	}
 }